refactor: extract worker-host, vscode-request, session-totals, and loading-grid modules
Behavior-preserving, test-gated readability pass for issue #106. Splits the largest modules (parser.ts, parser-vscode.ts, app.ts) into focused units and removes rule-of-three duplication. +22 tests, no runtime behavior change.
