+++ cspell.json
@@ -4,6 +4,7 @@
   "minWordLength": 5,
   "words": [
     "aicoach",
+    "acked",
     "affordances",
     "allpending",
     "antipatterns",
@@ -13,6 +14,7 @@
     "browsable",
     "btcbb",
     "burndown",
+    "catchable",
     "chartjs",
     "chatmode",
     "chatmodes",
@@ -68,6 +70,7 @@
     "segoe",
     "sidechain",
     "sparkline",
+    "sysfree",
     "tamas",
     "todoread",
     "tokenless",
@@ -76,6 +79,8 @@
     "treemap",
     "tseslint",
     "undercount",
+    "unacked",
+    "unflushed",
     "unparseable",
     "upskilling",
     "visualbasic",
