+++ src/core/parser-vscode.test.ts
@@ -743,6 +743,71 @@ describe('parseSessionFile — skill detection', () => {
     });
   });
 });
+describe('tool-call metadata extraction (characterization)', () => {
+  function parseRequest(metadata: Record<string, unknown>): ReturnType<typeof parseSessionFile> {
+    const data = {
+      sessionId: 'tc',
+      requests: [{
+        requestId: 'r1',
+        timestamp: 1700000001000,
+        message: { text: 'go' },
+        response: [{ value: 'done' }],
+        result: { metadata },
+      }],
+    };
+    let out: ReturnType<typeof parseSessionFile> = null;
+    withTempFile('tc.json', JSON.stringify(data), (filePath) => {
+      out = parseSessionFile(filePath, 'ws', 'wp', 'Local Agent');
+    });
+    return out;
+  }
+
+  it('collects toolsUsed from toolCallRounds', () => {
+    const session = parseRequest({
+      toolCallRounds: [{ toolCalls: [{ name: 'read_file' }, { name: 'apply_patch' }] }],
+    });
+    expect(session!.requests[0].toolsUsed).toEqual(['read_file', 'apply_patch']);
+  });
+
+  it('collects toolsUsed from the toolCallResults branch too', () => {
+    const session = parseRequest({
+      toolCallResults: [{ toolCalls: [{ name: 'run_in_terminal' }] }],
+    });
+    expect(session!.requests[0].toolsUsed).toEqual(['run_in_terminal']);
+  });
+
+  it('parses toolCalls supplied as a JSON string', () => {
+    const session = parseRequest({
+      toolCallRounds: [{ toolCalls: JSON.stringify([{ name: 'edit_file' }]) }],
+    });
+    expect(session!.requests[0].toolsUsed).toEqual(['edit_file']);
+  });
+
+  it('skips malformed toolCalls JSON without throwing', () => {
+    const session = parseRequest({
+      toolCallRounds: [{ toolCalls: '{not valid json' }],
+    });
+    expect(session).not.toBeNull();
+    expect(session!.requests[0].toolsUsed).toEqual([]);
+  });
+
+  it('captures the last manage_todo_list snapshot', () => {
+    const session = parseRequest({
+      toolCallRounds: [
+        { toolCalls: [{ name: 'manage_todo_list', arguments: JSON.stringify({ todoList: [{ id: 1, title: 'first', status: 'completed' }] }) }] },
+        { toolCalls: [{ name: 'manage_todo_list', arguments: JSON.stringify({ todoList: [{ id: 2, title: 'second', status: 'in-progress' }] }) }] },
+      ],
+    });
+    expect(session!.requests[0].todoSnapshot).toEqual([
+      { id: 2, title: 'second', status: 'in-progress' },
+    ]);
+  });
+
+  it('returns null todoSnapshot when no manage_todo_list call is present', () => {
+    const session = parseRequest({ toolCallRounds: [{ toolCalls: [{ name: 'read_file' }] }] });
+    expect(session!.requests[0].todoSnapshot).toBeNull();
+  });
+});
 describe('harnessFromPath — VS Code Server', () => {
   it('returns "Local Agent (Server)" for .vscode-server paths', () => {
     expect(harnessFromPath('/home/alice/.vscode-server/data/User/workspaceStorage')).toBe('Local Agent (Server)');
